requirements.txt
Django
pytz==2018.5
python-dotenv
psycopg2-binary
djangorestframework
requests
django-rest-auth
django-allauth
django_rest_passwordreset
django_filter
